fix: Rename module & update references github.com/mumoshu/vals -> github.com/variantdev/vals
