+++ pkg/providers/sops/sops.go
@@ -2,7 +2,7 @@ package sops
 
 import (
 	"fmt"
-	"github.com/mumoshu/vals/pkg/api"
+	"github.com/variantdev/vals/pkg/api"
 	"gopkg.in/yaml.v3"
 	"os"
 	"strings"
