+++ pkg/stringmapprovider/stringmapprovider.go
@@ -2,11 +2,11 @@ package stringmapprovider
 
 import (
 	"fmt"
-	"github.com/mumoshu/vals/pkg/api"
-	"github.com/mumoshu/vals/pkg/providers/awssecrets"
-	"github.com/mumoshu/vals/pkg/providers/sops"
-	"github.com/mumoshu/vals/pkg/providers/ssm"
-	"github.com/mumoshu/vals/pkg/providers/vault"
+	"github.com/variantdev/vals/pkg/api"
+	"github.com/variantdev/vals/pkg/providers/awssecrets"
+	"github.com/variantdev/vals/pkg/providers/sops"
+	"github.com/variantdev/vals/pkg/providers/ssm"
+	"github.com/variantdev/vals/pkg/providers/vault"
 )
 
 func New(provider api.StaticConfig) (api.LazyLoadedStringMapProvider, error) {
