+++ vals_awssecrets_test.go
@@ -1,73 +1,19 @@
-package values
+package vals
 
 import (
 	"fmt"
 	"testing"
 )
 
-func TestValues_Vault_EvalTemplate(t *testing.T) {
+func TestValues_AWSSecrets_String(t *testing.T) {
 	// TODO
 	// Pre-requisite:
-	//   vault secrets enable -path=mykv kv
-	//   vault write mykv/foo mykey=myvalue
-	//   vault read mykv/foo
+	//   aws secretsmanager create-secret --name /mykv/foo/mykey --secret-string myvalue
 
 	type testcase struct {
 		config map[string]interface{}
 	}
 
-	testcases := []testcase{
-		{
-			config: map[string]interface{}{
-				"foo": "ref+vault://127.0.0.1:8200/mykv/foo?proto=http#/mykey",
-				"bar": map[string]interface{}{
-					"baz": "ref+vault://127.0.0.1:8200/mykv/foo?proto=http#/mykey",
-				},
-			},
-		},
-	}
-
-	for i := range testcases {
-		tc := testcases[i]
-		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
-			vals, err := Eval(tc.config)
-			if err != nil {
-				t.Fatalf("%v", err)
-			}
-
-			{
-				expected := "myvalue"
-				key := "foo"
-				actual := vals[key]
-				if actual != expected {
-					t.Errorf("unepected value for key %q: expected=%q, got=%q", key, expected, actual)
-				}
-			}
-
-			{
-				switch bar := vals["bar"].(type) {
-				case map[string]interface{}:
-					expected := "myvalue"
-					key := "baz"
-					actual := bar[key]
-					if actual != expected {
-						t.Errorf("unepected value for key %q: expected=%q, got=%q", key, expected, actual)
-					}
-				default:
-					t.Fatalf("unexpected type of bar: value=%v, type=%T", bar, bar)
-				}
-			}
-		})
-	}
-}
-
-func TestValues_Vault_String(t *testing.T) {
-	// TODO
-	// Pre-requisite: vault write mykv/foo mykey=myvalue
-
-	type testcase struct {
-		config map[string]interface{}
-	}
 	commonInline := map[string]interface{}{
 		"foo": "mykey",
 		"bar": map[string]interface{}{
@@ -79,10 +25,10 @@ func TestValues_Vault_String(t *testing.T) {
 		{
 			config: map[string]interface{}{
 				"provider": map[string]interface{}{
-					"name":    "vault",
-					"type":    "string",
-					"path":    "mykv/foo",
-					"address": "http://127.0.0.1:8200",
+					"name":   "awssecrets",
+					"type":   "string",
+					"path":   "/mykv/foo",
+					"region": "ap-northeast-1",
 				},
 				"inline": commonInline,
 			},
@@ -90,20 +36,20 @@ func TestValues_Vault_String(t *testing.T) {
 		{
 			config: map[string]interface{}{
 				"provider": map[string]interface{}{
-					"name": "vault",
+					"name": "awssecrets",
 					// implies type=string
-					"path":    "mykv/foo",
-					"address": "http://127.0.0.1:8200",
+					"path":   "/mykv/foo",
+					"region": "ap-northeast-1",
 				},
 				"inline": commonInline,
 			},
 		},
 		{
 			config: map[string]interface{}{
-				// implies name=vault and type=string
-				"vault": map[string]interface{}{
-					"path":    "mykv/foo",
-					"address": "http://127.0.0.1:8200",
+				// implies name=ssm and type=string
+				"awssecrets": map[string]interface{}{
+					"path":   "/mykv/foo",
+					"region": "ap-northeast-1",
 				},
 				"inline": commonInline,
 			},
@@ -146,140 +92,56 @@ func TestValues_Vault_String(t *testing.T) {
 	}
 }
 
-func TestValues_Vault_Map(t *testing.T) {
+func TestValues_AWSSecrets_Map(t *testing.T) {
 	// TODO
-	// Pre-requisite: vault write mykv/foo mykey=myvalue
+	// Pre-requisite:
+	//   aws secretsmanager create-secret --name /mykv/foo/meta --secret-string '{"github.com/variantdev/vals":["mykey"]}'
+	//   aws secretsmanager create-secret --name /mykv/foo/mykey --secret-string myvalue
 
 	type testcase struct {
-		name   string
-		config map[string]interface{}
+		provider map[string]interface{}
 	}
+
 	testcases := []testcase{
 		{
-			config: map[string]interface{}{
-				"inline": map[string]interface{}{
-					"foo": "foo",
-					"bar": map[string]interface{}{
-						"baz": "foo",
-					},
-				},
-				"provider": map[string]interface{}{
-					"name":    "vault",
-					"type":    "map",
-					"path":    "mykv",
-					"address": "http://127.0.0.1:8200",
-				},
+			provider: map[string]interface{}{
+				"name":   "awssecrets",
+				"type":   "map",
+				"path":   "/mykv",
+				"region": "ap-northeast-1",
 			},
 		},
 		{
-			config: map[string]interface{}{
-				"inline": map[string]interface{}{
-					"foo": "foo",
-					"bar": map[string]interface{}{
-						"baz": "foo",
-					},
-				},
-				"provider": map[string]interface{}{
-					"name":    "vault",
-					"type":    "map",
-					"format":  "raw",
-					"path":    "mykv",
-					"address": "http://127.0.0.1:8200",
-				},
+			provider: map[string]interface{}{
+				"name":   "awssecrets",
+				"type":   "map",
+				"format": "raw",
+				"path":   "/mykv",
+				"region": "ap-northeast-1",
 			},
 		},
 		{
-			config: map[string]interface{}{
-				"inline": map[string]interface{}{
-					"foo": "foo",
-					"bar": map[string]interface{}{
-						"baz": "foo",
-					},
-				},
-				"provider": map[string]interface{}{
-					"name": "vault",
-					// implies type:map format:raw
-					"prefix":  "mykv",
-					"address": "http://127.0.0.1:8200",
-				},
-			},
-		},
-		{
-			name: "setForKey1",
-			config: map[string]interface{}{
-				"vault": map[string]interface{}{
-					// implies type:map format:raw
-					"prefix":     "mykv/foo",
-					"address":    "http://127.0.0.1:8200",
-					"setForKeys": []string{"foo", "bar.baz"},
-				},
-			},
-		},
-		{
-			name: "setForKey2",
-			config: map[string]interface{}{
-				"vault": map[string]interface{}{
-					// implies type:map format:raw
-					"paths":      []string{"mykv/foo/mykey"},
-					"address":    "http://127.0.0.1:8200",
-					"setForKeys": []string{"foo", "bar.baz"},
-				},
-			},
-		},
-		{
-			name: "setForKey3",
-			config: map[string]interface{}{
-				"vault": map[string]interface{}{
-					// implies type:map format:raw
-					"prefix":     "mykv/foo/",
-					"keys":       []string{"mykey"},
-					"address":    "http://127.0.0.1:8200",
-					"setForKeys": []string{"foo", "bar.baz"},
-				},
-			},
-		},
-		{
-			name: "set1",
-			config: map[string]interface{}{
-				"vault": map[string]interface{}{
-					// implies type:map format:raw
-					"prefix":  "mykv",
-					"address": "http://127.0.0.1:8200",
-					"set": map[string]interface{}{
-						"foo": "foo",
-						"bar": map[string]interface{}{
-							"baz": "foo",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "set2",
-			config: map[string]interface{}{
-				"vault": map[string]interface{}{
-					// implies type:map format:raw
-					"prefix":  "mykv",
-					"address": "http://127.0.0.1:8200",
-					"set": map[string]interface{}{
-						"foo": "foo",
-						"bar": map[string]interface{}{
-							"baz": "foo",
-						},
-					},
-				},
+			provider: map[string]interface{}{
+				"name": "awssecrets",
+				// implies type:map format:raw
+				"prefix": "/mykv",
+				"region": "ap-northeast-1",
 			},
 		},
 	}
 
 	for i := range testcases {
 		tc := testcases[i]
-		tcname := fmt.Sprintf("%d", i)
-		if tc.name != "" {
-			tcname = tc.name
-		}
-		t.Run(tcname, func(t *testing.T) {
-			config := Map(tc.config)
+		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
+			config := Map(map[string]interface{}{
+				"provider": tc.provider,
+				"inline": map[string]interface{}{
+					"foo": "foo",
+					"bar": map[string]interface{}{
+						"baz": "foo",
+					},
+				},
+			})
 
 			vals, err := Load(config)
 			if err != nil {
@@ -328,30 +190,31 @@ func TestValues_Vault_Map(t *testing.T) {
 	}
 }
 
-func TestValues_Vault_Map_Raw(t *testing.T) {
+func TestValues_AWSSecrets_Map_Raw(t *testing.T) {
 	// TODO
-	// Pre-requisite: vault write mykv/foo mykey=myvalue
+	// Pre-requisite: aws ssm put-parameter --name /mykv/foo/mykey --value myvalue --type String
 
 	type testcase struct {
 		provider map[string]interface{}
 	}
+
 	testcases := []testcase{
 		{
 			provider: map[string]interface{}{
-				"name":    "vault",
+				"name":    "awssecrets",
 				"type":    "map",
-				"path":    "mykv",
+				"path":    "/mykv",
 				"address": "http://127.0.0.1:8200",
 				"format":  "raw",
 			},
 		},
 		{
 			provider: map[string]interface{}{
-				"name": "vault",
+				"name": "awssecrets",
 				// implies
 				//"type":    "map",
 				//"format":  "raw",
-				"prefix":  "mykv",
+				"prefix":  "/mykv",
 				"address": "http://127.0.0.1:8200",
 			},
 		},
@@ -418,7 +281,7 @@ func TestValues_Vault_Map_Raw(t *testing.T) {
 	}
 }
 
-func TestValues_Vault_Map_YAML(t *testing.T) {
+func TestValues_AWSSecrets_Map_YAML(t *testing.T) {
 	// TODO
 	// cat <<EOF > myyaml.yaml
 	// baz:
@@ -429,26 +292,29 @@ func TestValues_Vault_Map_YAML(t *testing.T) {
 	// {"baz": {"mykey": "myvalue"}}
 	// EOF
 	//
-	// vault write mykv/yamltest myyaml="$(cat myyaml.yaml)" myjson="$(cat myjson.json)"
+	// Pre-requisite:
+	//   aws secretsmanager create-secret --name /mykv/yamltest/myyaml --secret-string "$(cat myyaml.yaml)"
+	//   aws secretsmanager create-secret --name /mykv/yamltest/myjson --secret-string "$(cat myjson.json)"
 
 	type testcase struct {
 		provider map[string]interface{}
 		dataKey  string
 	}
+
 	provider1 := map[string]interface{}{
-		"name":    "vault",
-		"type":    "map",
-		"path":    "mykv/yamltest",
-		"address": "http://127.0.0.1:8200",
-		"format":  "yaml",
+		"name":   "awssecrets",
+		"type":   "map",
+		"path":   "/mykv/yamltest",
+		"region": "ap-northeast-1",
+		"format": "yaml",
 	}
 
 	provider2 := map[string]interface{}{
-		"name": "vault",
+		"name": "awssecrets",
 		// implies `type: map`
-		"path":    "mykv/yamltest",
-		"address": "http://127.0.0.1:8200",
-		"format":  "yaml",
+		"path":   "/mykv/yamltest",
+		"region": "ap-northeast-1",
+		"format": "yaml",
 	}
 
 	testcases := []testcase{
@@ -510,7 +376,7 @@ func TestValues_Vault_Map_YAML(t *testing.T) {
 	}
 }
 
-func TestValues_Vault_Map_YAML_Root(t *testing.T) {
+func TestValues_AWSSecrets_Map_YAML_Root(t *testing.T) {
 	// TODO
 	// cat <<EOF > myyaml.yaml
 	// baz:
@@ -521,7 +387,9 @@ func TestValues_Vault_Map_YAML_Root(t *testing.T) {
 	// {"baz": {"mykey": "myvalue"}}
 	// EOF
 	//
-	// vault write mykv/yamltest myyaml="$(cat myyaml.yaml)" myjson="$(cat myjson.json)"
+	// Pre-requisite:
+	//   aws secretsmanager create-secret --name /mykv/yamltest/myyaml --secret-string "$(cat myyaml.yaml)"
+	//   aws secretsmanager create-secret --name /mykv/yamltest/myjson --secret-string "$(cat myjson.json)"
 
 	type provider struct {
 		config map[string]interface{}
@@ -530,52 +398,52 @@ func TestValues_Vault_Map_YAML_Root(t *testing.T) {
 		{
 			config: map[string]interface{}{
 				"provider": map[string]interface{}{
-					"name":    "vault",
-					"type":    "map",
-					"path":    "mykv/yamltest/myyaml",
-					"address": "http://127.0.0.1:8200",
-					"format":  "yaml",
+					"name":   "awssecrets",
+					"type":   "map",
+					"path":   "/mykv/yamltest/myyaml",
+					"region": "ap-northeast-1",
+					"format": "yaml",
 				},
 			},
 		},
 		{
 			config: map[string]interface{}{
 				"provider": map[string]interface{}{
-					"name":    "vault",
-					"type":    "map",
-					"path":    "mykv/yamltest/myjson",
-					"address": "http://127.0.0.1:8200",
-					"format":  "yaml",
+					"name":   "awssecrets",
+					"type":   "map",
+					"path":   "/mykv/yamltest/myjson",
+					"region": "ap-northeast-1",
+					"format": "yaml",
 				},
 			},
 		},
 		{
 			config: map[string]interface{}{
 				"provider": map[string]interface{}{
-					"name": "vault",
+					"name": "awssecrets",
 					// implies format:yaml and type:map
-					"path":    "mykv/yamltest/myyaml",
-					"address": "http://127.0.0.1:8200",
+					"path":   "/mykv/yamltest/myyaml",
+					"region": "ap-northeast-1",
 				},
 			},
 		},
 		{
 			config: map[string]interface{}{
 				"provider": map[string]interface{}{
-					"name": "vault",
+					"name": "awssecrets",
 					// implies format:yaml and type:map
-					"path":    "mykv/yamltest/myjson",
-					"address": "http://127.0.0.1:8200",
+					"path":   "/mykv/yamltest/myjson",
+					"region": "ap-northeast-1",
 				},
 			},
 		},
 		{
 			config: map[string]interface{}{
 				// implies name:vault
-				"vault": map[string]interface{}{
+				"awssecrets": map[string]interface{}{
 					// implies format:yaml and type:map
-					"path":    "mykv/yamltest/myjson",
-					"address": "http://127.0.0.1:8200",
+					"path":   "/mykv/yamltest/myjson",
+					"region": "ap-northeast-1",
 				},
 			},
 		},
@@ -611,7 +479,7 @@ func TestValues_Vault_Map_YAML_Root(t *testing.T) {
 	}
 }
 
-func TestValues_Vault_Map_Raw_Root(t *testing.T) {
+func TestValues_AWSSecrets_Map_Raw_Root(t *testing.T) {
 	// TODO
 	// cat <<EOF > myyaml.yaml
 	// baz:
@@ -622,24 +490,26 @@ func TestValues_Vault_Map_Raw_Root(t *testing.T) {
 	// {"baz": {"mykey": "myvalue"}}
 	// EOF
 	//
-	// vault write mykv/yamltest myyaml="$(cat myyaml.yaml)" myjson="$(cat myjson.json)"
+	// Pre-requisite:
+	//   aws secretsmanager create-secret --name /mykv/foo/mykey --secret-string myvalue
 
 	type testcase struct {
 		config map[string]interface{}
 	}
+
 	provider1 := map[string]interface{}{
-		"name":    "vault",
-		"type":    "map",
-		"path":    "mykv/foo",
-		"address": "http://127.0.0.1:8200",
-		"format":  "raw",
+		"name":   "awssecrets",
+		"type":   "map",
+		"path":   "/mykv/foo",
+		"region": "ap-northeast-1",
+		"format": "raw",
 	}
 
 	provider2 := map[string]interface{}{
-		"name": "vault",
+		"name": "awssecrets",
 		// implies format:raw
-		"prefix":  "mykv/foo",
-		"address": "http://127.0.0.1:8200",
+		"prefix": "/mykv/foo",
+		"region": "ap-northeast-1",
 	}
 
 	testcases := []testcase{
@@ -655,21 +525,10 @@ func TestValues_Vault_Map_Raw_Root(t *testing.T) {
 		},
 		{
 			config: map[string]interface{}{
-				// implies name:vault
-				"vault": map[string]interface{}{
-					// implies format:raw
-					"prefix":  "mykv/foo",
-					"address": "http://127.0.0.1:8200",
-				},
-			},
-		},
-		{
-			config: map[string]interface{}{
 				// implies name:ssm
-				"vault": map[string]interface{}{
+				"awssecrets": map[string]interface{}{
 					// implies format:raw
 					"prefix":  "/mykv/foo",
-					"keys":    []string{"mykey"},
 					"address": "http://127.0.0.1:8200",
 				},
 			},
@@ -677,10 +536,21 @@ func TestValues_Vault_Map_Raw_Root(t *testing.T) {
 		{
 			config: map[string]interface{}{
 				// implies name:ssm
-				"vault": map[string]interface{}{
+				"awssecrets": map[string]interface{}{
 					// implies format:raw
-					"paths":   []string{"/mykv/foo/mykey"},
-					"address": "http://127.0.0.1:8200",
+					"prefix": "/mykv/foo",
+					"keys":   []string{"mykey"},
+					"region": "ap-northeast-1",
+				},
+			},
+		},
+		{
+			config: map[string]interface{}{
+				// implies name:ssm
+				"awssecrets": map[string]interface{}{
+					// implies format:raw
+					"paths":  []string{"/mykv/foo/mykey"},
+					"region": "ap-northeast-1",
 				},
 			},
 		},
