+++ vals_ssm_test.go
@@ -1,4 +1,4 @@
-package values
+package vals
 
 import (
 	"fmt"
